- Added module activation - Added test damaging ability

+++ BattleDroids/Assets/Scripts/GameObjects/Destructible.cs
@@ -5,9 +5,9 @@ using UnityEngine;
 public class Destructible : ScriptableObject
 {
     [SerializeField]
-    protected int m_durability = 100, m_integrity = 100;
+    protected float m_durability = 100.0f, m_integrity = 100.0f;
 
-    public int Damage(int _amount)
+    public float Damage(float _amount)
     {
         m_integrity -= _amount;
 
@@ -19,7 +19,7 @@ public class Destructible : ScriptableObject
         return m_integrity;
     }
 
-    public int Repair(int _amount)
+    public float Repair(float _amount)
     {
         m_integrity += _amount;
 
@@ -31,7 +31,7 @@ public class Destructible : ScriptableObject
         return m_integrity;
     }
 
-    public int RepairFull()
+    public float RepairFull()
     {
         m_integrity = m_durability;
 
@@ -40,25 +40,25 @@ public class Destructible : ScriptableObject
 
     public float GetIntegrityPercent()
     {
-        return (float)m_integrity / (float)m_durability;
+        return m_integrity / m_durability;
     }
 
-    public int GetDurability()
+    public float GetDurability()
     {
         return m_durability;
     }
 
-    public int GetIntegrity()
+    public float GetIntegrity()
     {
         return m_integrity;
     }
 
-    public void SetDurability(int _durability)
+    public void SetDurability(float _durability)
     {
         m_durability = _durability;
     }
 
-    public void SetIntegrity(int _integrity)
+    public void SetIntegrity(float _integrity)
     {
         m_integrity = _integrity;
     }
